requirements.txt
# Production dependencies
Django>=5.0,<6.0
gunicorn>=21.2
whitenoise>=6.6
dj-database-url>=2.1
python-dotenv>=1.0
psycopg[binary]>=3.1
qrcode[pil]>=7.4

# Frontend & dev-only in production? Уточним ниже
django-tailwind
django-browser-reload